Connexion v1

+++ User/Delete.php
@@ -1,27 +1,40 @@
-<?php
-	include "../General/connectionBD.php";
+<?php include "../General/isAdmin.php" //$IsAdmin == true si Admin ?>
 
-	try {
-		$DB->beginTransaction();
-		$insert = $DB->prepare("DELETE FROM USER WHERE Login = :Login  ");
-		$data = array(
-			':Login'=>$_POST["Login"],
+<?php if ($isAdmin) { ?>
 
-		);
+	<?php
+		include "../General/connectionBD.php";
+
+		try {
+			$DB->beginTransaction();
+			$insert = $DB->prepare("DELETE FROM USER WHERE Login = :Login  ");
+			$data = array(
+				':Login'=>$_POST["Login"],
+
+			);
+			
+			$insert->execute($data);
+			$DB->commit();
+
+
+
+		}
+		catch(PDOException $e){
+			echo "<p> You can't Delete that ! </p>", "<p>", $e , "</p>";
+			$DB->rollBack();
+		}
+
+		header('Location: ./AllUser.php');
+		exit();
 		
-		$insert->execute($data);
-		$DB->commit();
 
+	?>
 
-
-	}
-	catch(PDOException $e){
-		echo "<p> You can't Delete that ! </p>", "<p>", $e , "</p>";
-		$DB->rollBack();
-	}
-
-	header('Location: ./AllUser.php');
+<?php 
+}
+else{
+	header('Location: ../index.php');
 	exit();
-	
+	} 
 
- ?>
+?>
